+++ src/gui/mainwindow.h
@@ -54,6 +54,7 @@ private:
     SettingsTab      *m_settingsTab  = nullptr;
     QTextEdit        *m_logEdit      = nullptr;
     int               m_logTabIndex  = -1;
+    quint64           m_lastRuleFailureGen = 0;
     QCheckBox        *m_logAutoScroll = nullptr;
     QLabel           *m_tempStatus   = nullptr;
     QWidget          *m_safeBanner   = nullptr;